+++ src/content/popup/show-popup.ts
@@ -70,6 +70,7 @@ export type ShowPopupOptions = {
   posDisplay: PartOfSpeechDisplay;
   positionMode: PopupPositionMode;
   popupStyle: string;
+  preferredUnits: 'metric' | 'imperial';
   previousHeight?: number;
   safeArea: Box;
   showDefinitions: boolean;